TreeRoot now build as a static library & Fixed some compiler warnings & Integrated vendor premake files together.

+++ TreeRoot/src/Platform/Windows/WindowsWindow.cpp
@@ -135,7 +135,7 @@ namespace tr
 		{
 			WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
 
-			MouseMovedEvent e(xpos, ypos);
+			MouseMovedEvent e((float)xpos, (float)ypos);
 			data.EventCallback(e);
 		});
 
@@ -143,7 +143,7 @@ namespace tr
 		{
 			WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
 
-			MouseScrolledEvent e(xoffset, yoffset);
+			MouseScrolledEvent e((float)xoffset, (float)yoffset);
 			data.EventCallback(e);
 		});
 	}
